fix: resolve all clippy warnings across workspace
Auto-fixed (cargo clippy --fix):
- unnecessary u32→u32 cast (agentic_loop_turn.rs)
- map_or simplification (command_registry.rs, repl_ui.rs)
- bool::then closure → bool::then_some (repl_turn.rs)
- iter().cloned().collect() → to_vec() (slash_team.rs)
- explicit copy closure → copied() (slash_tuning.rs)
- collapsible if statements (session_journal.rs, session_workspace.rs)
- or_insert_with(T::new) → or_default() (user_profile.rs)
- ref-in-ref pattern (delegation_engine.rs)
- redundant let binding (main.rs)
- unnecessary borrow (main.rs)

Manual fixes:
- format! in format! args → inline {:?} (decision_explainer.rs)
- manual clamp → .clamp() (goal_tracker.rs)
- complex tuple type → LspTextEdit alias (lsp_tools.rs)
- unwrap after is_none → if-let-some (slash_session.rs)
- div_ceil reimpl → .div_ceil() (system.rs)
- is_multiple_of reimpl → .is_multiple_of() (ab_testing.rs)

Suppressed (intentional design):
- #[allow(clippy::should_implement_trait)] on SubRunState::from_str
- #[allow(clippy::new_ret_no_self)] on Experiment::new → builder

+++ rust/crates/astra-cli/src/cli/chat_stream/sse_loop/agentic_loop_turn.rs
@@ -413,7 +413,7 @@ async fn prepare_chat_turn_payload(ctx: PrepareChatTurnRequest<'_>) -> Value {
 
     // ─── Record token budget estimate to trace collector (M1 observability) ───
     if let Some(collector) = ctx.telem.trace_collector {
-        let schema_tokens = ctx.selector.registry().total_pinned_token_cost() as u32;
+        let schema_tokens = ctx.selector.registry().total_pinned_token_cost();
         let budget = prompts::budget_for_model(ctx.model);
         let max_tokens = budget.model_limit as u32;
